refactor: Consolidate result handling and update import paths
- Replaced references to obsolete error and success handlers with a new `result-handlers` module, improving code organization and maintainability.
- Updated import paths across multiple scripts to reflect the new structure, ensuring consistent error and success handling throughout the application.
- Removed outdated success handler implementations, streamlining the success logging process for better clarity and consistency.
